resolve: preserve dev version of pagination for review and keep the new one
